chore(snapshot): capture DSP reliability and backend work
Snapshot the current CUDA/Triton DSP lifecycle fixes together with the Vulkan, SDX AOT, native-image, and backend test work.

Validated the DSP core gate at 1643/0/0 twice and the MTP oracle at 60/60.

+++ libnd4j/include/ops/declarable/helpers/roll.h
@@ -26,7 +26,7 @@ namespace sd {
 namespace ops {
 namespace helpers {
 
-SD_LIB_HIDDEN void rollFunctorLinear(LaunchContext* context, NDArray* input, NDArray* output, int shift,
+SD_LIB_HIDDEN void rollFunctorLinear(LaunchContext* context, NDArray* input, NDArray* output, LongType shift,
                                      bool inplace = false);
 
 SD_LIB_HIDDEN void rollFunctorFull(LaunchContext* context, NDArray* input, NDArray* output,
